requirements.txt
Django==2.1.2
django-rest-multiple-models==2.1.1
djangorestframework==3.8.2
django-phonenumber-field==1.3.0
drf-openapi==1.3.0
simplejson==3.16.0
django-filter
openpyxl
